# Conflicts:
#	src/main/java/com/sparta/travelshooting/reviewPost/controller/ReviewPostController.java
#	src/main/java/com/sparta/travelshooting/reviewPost/service/ReviewPostServiceImpl.java
#	src/main/resources/templates/home.html

+++ src/main/java/com/sparta/travelshooting/reviewPost/controller/ReviewPostController.java
@@ -5,11 +5,14 @@ import com.sparta.travelshooting.reviewPost.dto.HomeReviewResponseDto;
 import com.sparta.travelshooting.reviewPost.dto.ReviewPostListResponseDto;
 import com.sparta.travelshooting.reviewPost.dto.ReviewPostRequestDto;
 import com.sparta.travelshooting.reviewPost.dto.ReviewPostResponseDto;
+import com.sparta.travelshooting.reviewPost.entity.ReviewPost;
 import com.sparta.travelshooting.reviewPost.service.ReviewPostService;
 import com.sparta.travelshooting.security.UserDetailsImpl;
+import com.sparta.travelshooting.user.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -21,10 +24,13 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.RejectedExecutionException;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/api/review-posts")
+@Slf4j
 @Tag(name = "여행 계획 후기 게시글 API")
 public class ReviewPostController {
 
@@ -58,6 +64,11 @@ public class ReviewPostController {
     @Operation(summary = "후기 게시글 삭제")
     @DeleteMapping("/{reviewPostId}")
     public ResponseEntity<ApiResponseDto> deleteReviewPost(@PathVariable Long reviewPostId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        // 로그인이 되어있지 않은 경우 처리
+        if (userDetails == null) {
+            return new ResponseEntity<>(new ApiResponseDto("로그인이 필요합니다.", HttpStatus.UNAUTHORIZED.value()), HttpStatus.OK);
+        }
+
         ApiResponseDto apiResponseDto = reviewPostService.deleteReviewPost(reviewPostId, userDetails.getUser());
         return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
     }
@@ -108,19 +119,30 @@ public class ReviewPostController {
     //좋아요 여부 조회
     @GetMapping("/like/{reviewPostId}")
     public ResponseEntity<ApiResponseDto> checkLikeStatus(@PathVariable Long reviewPostId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        // 로그인이 되어있지 않은 경우 처리
+        if (userDetails == null) {
+            return new ResponseEntity<>(new ApiResponseDto("로그인이 필요합니다.", HttpStatus.UNAUTHORIZED.value()), HttpStatus.OK);
+        }
+
         Long userId = userDetails.getUser().getId();
         boolean isLiked = reviewPostService.hasLiked(reviewPostId, userId);
-        if (isLiked) {return new ResponseEntity<>(new ApiResponseDto("이미 좋아요를 누른 상태입니다.", 200), HttpStatus.OK);
-        } else {return new ResponseEntity<>(new ApiResponseDto("아직 좋아요를 누르지 않은 상태입니다.", 200), HttpStatus.OK);}
+        if (isLiked) {
+            return new ResponseEntity<>(new ApiResponseDto("이미 좋아요를 누른 상태입니다.", 200), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(new ApiResponseDto("아직 좋아요를 누르지 않은 상태입니다.", 200), HttpStatus.OK);
+        }
     }
 
 
     //작성자 확인
     @GetMapping("/check-user/{reviewPostId}")
     public ResponseEntity<Boolean> reviewPostCheckUser(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long reviewPostId) {
+        //로그인 되어있지 않은 경우 처리
+        if (userDetails == null) {
+            return ResponseEntity.ok(false);
+        }
+
         boolean isAuthor = reviewPostService.reviewPostCheckUser(userDetails, reviewPostId);
-        return ResponseEntity.ok(isAuthor);}
-
-
-
+        return ResponseEntity.ok(isAuthor);
+    }
 }
